#296 Rename roboconf-dm-monitoring in roboconf-dm-templating
