+++ core/roboconf-agent/src/main/java/net/roboconf/agent/internal/AgentMessageProcessor.java
@@ -223,7 +223,6 @@ public class AgentMessageProcessor extends AbstractMessageProcessor<IAgentClient
 
 			this.agent.setScopedInstance( newScopedInstance );
 			instancesToProcess.addAll( InstanceHelpers.buildHierarchicalList( this.scopedInstance ));
-			instancesToProcess.remove( this.scopedInstance );
 
 			// Notify the DM
 			if( this.scopedInstance.getStatus() != InstanceStatus.DEPLOYED_STARTED ) {
@@ -237,8 +236,6 @@ public class AgentMessageProcessor extends AbstractMessageProcessor<IAgentClient
 		}
 
 		// Configure the messaging.
-		// The scoped instance is excluded from this part as it does not make
-		// sense for it to have dependencies.
 		for( Instance instanceToProcess : instancesToProcess ) {
 			this.messagingClient.listenToExportsFromOtherAgents( ListenerCommand.START, instanceToProcess );
 			this.messagingClient.requestExportsFromOtherAgents( instanceToProcess );
