--- core/roboconf-dm-monitoring/src/main/java/net/roboconf/dm/monitoring/internal/ApplicationContextBean.java
@@ -1,76 +1,0 @@
-/**
- * Copyright 2013-2015 Linagora, Université Joseph Fourier, Floralis
- *
- * The present code is developed in the scope of the joint LINAGORA -
- * Université Joseph Fourier - Floralis research program and is designated
- * as a "Result" pursuant to the terms and conditions of the LINAGORA
- * - Université Joseph Fourier - Floralis research program. Each copyright
- * holder of Results enumerated here above fully & independently holds complete
- * ownership of the complete Intellectual Property rights applicable to the whole
- * of said Results, and may freely exploit it in any manner which does not infringe
- * the moral rights of the other copyright holders.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.roboconf.dm.monitoring.internal;
-
-import java.util.Collection;
-import java.util.Date;
-import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
-import java.util.Map;
-import java.util.Set;
-
-/**
- * Context bean for a Roboconf application.
- *
- * @author Pierre Bourret - Université Joseph Fourier
- */
-public class ApplicationContextBean {
-	String name;
-	String description;
-	Date lastModified;
-	final Set<InstanceContextBean> instances = new LinkedHashSet<InstanceContextBean>();
-	final Set<String> components = new LinkedHashSet<String>();
-	final Map<String, Set<InstanceContextBean>> instancesByType = new LinkedHashMap<String, Set<InstanceContextBean>>();
-
-	public String getName() {
-		return this.name;
-	}
-
-	public String getDescription() {
-		return description;
-	}
-
-	public Date getLastModified() {
-		return this.lastModified;
-	}
-
-	public Collection<InstanceContextBean> getInstances() {
-		return this.instances;
-	}
-
-	public Collection<String> getComponents() {
-		return this.components;
-	}
-
-	public Map<String, Set<InstanceContextBean>> getInstancesByType() {
-		return this.instancesByType;
-	}
-
-	@Override
-	public String toString() {
-		return this.name;
-	}
-}
