+++ core/roboconf-dm-templating/.gitignore
