+++ core/roboconf-dm-templating/src/main/resources/OSGI-INF/bundle.info
@@ -1,0 +1,7 @@
+\u001B[1mSYNOPSIS\u001B[0m
+    This bundle is part of the Roboconf project.
+
+\u001B[1mDESCRIPTION\u001B[0m
+    The DM (Deployment Manager) templating generates reports driven by
+    user-defined templates. It watches the Roboconf configuration directory for templates
+    and generates and updates reports when the state of an application changes.
