+++ karaf/roboconf-karaf-feature-dm/src/main/feature/feature.xml
@@ -22,7 +22,7 @@
 		<bundle>mvn:net.roboconf/roboconf-dm/${project.version}</bundle>
 		<bundle>mvn:net.roboconf/roboconf-dm-web-administration/${project.version}/war</bundle>
 		<bundle>mvn:net.roboconf/roboconf-target-api/${project.version}</bundle>
-		<bundle>mvn:net.roboconf/roboconf-dm-monitoring/${project.version}</bundle>
+		<bundle>mvn:net.roboconf/roboconf-dm-templating/${project.version}</bundle>
 		
 		<!-- Optional bundles, to start depending on our needs -->
 		<!-- Deployment targets handlers could go here. -->
